requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Testing==0.4.2
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mock==1.0.1
more-itertools==4.1.0
nose==1.3.7
pbp.recipe.noserunner==0.2.6
pluggy==0.6.0
py==1.5.3
pytest==3.5.0
requests==2.18.4
six==1.11.0
slacker==0.9.65
tox==3.0.0
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.11.4
zc.buildout==2.11.3
zc.recipe.egg==2.0.3
python-gitlab==1.5.1
pyaml